# Reduce string allocations for literals [skip CI]