… .pre-commit-config.yaml
… hilti/runtime/src/tests/string.cc
… hilti/toolchain/src/compiler/codegen/ctors.cc
… hilti/toolchain/src/compiler/parser/parser.yy
… hilti/toolchain/src/compiler/visitors/normalizer.cc
… spicy/runtime/include/parser.h
… spicy/runtime/include/sink.h
… spicy/runtime/src/init.cc
… spicy/runtime/src/parser.cc
… spicy/runtime/src/sink.cc
… spicy/runtime/src/tests/init.cc
… spicy/toolchain/include/compiler/detail/codegen/parser-builder.h
… spicy/toolchain/src/compiler/codegen/codegen.cc
… spicy/toolchain/src/compiler/codegen/parser-builder.cc
… spicy/toolchain/src/compiler/codegen/unit-builder.cc
… spicy/toolchain/src/compiler/parser/parser.yy
… tests/Baseline/hilti.ast.coercion/output
… tests/Baseline/spicy.optimization.default-parser-functions/noopt.hlt
… tests/Baseline/spicy.optimization.default-parser-functions/opt.hlt
… tests/Baseline/spicy.optimization.feature_requirements/noopt.hlt
… tests/Baseline/spicy.optimization.feature_requirements/opt.hlt
… tests/Baseline/spicy.optimization.unused-functions/noopt.hlt
… tests/Baseline/spicy.optimization.unused-functions/opt.hlt
… tests/Baseline/spicy.optimization.unused-types/noopt.hlt
… tests/Baseline/spicy.optimization.unused-types/opt.hlt