With this PR the benchmark from #1589 shows identical performance for both cases for me. Overall runtime seems to improve as well.

Before:

```console
$ hyperfine --export-markdown results.md -w3 'cat test-data.txt | ./build/bin/spicy-driver t.hlto' 'cat test-data.txt | ./build/bin/spicy-driver t-compiled-with-realpath.hlto'
Benchmark 1: cat test-data.txt | ./build/bin/spicy-driver t.hlto
  Time (mean ± σ):      1.713 s ±  0.018 s    [User: 1.699 s, System: 0.031 s]
  Range (min … max):    1.695 s …  1.738 s    10 runs
 
Benchmark 2: cat test-data.txt | ./build/bin/spicy-driver t-compiled-with-realpath.hlto
  Time (mean ± σ):      2.166 s ±  0.014 s    [User: 2.141 s, System: 0.041 s]
  Range (min … max):    2.147 s …  2.191 s    10 runs
 
Summary
  cat test-data.txt | ./build/bin/spicy-driver t.hlto ran
    1.26 ± 0.02 times faster than cat test-data.txt | ./build/bin/spicy-driver t-compiled-with-realpath.hlto
```

After:

```console
$ hyperfine --export-markdown results.md -w3 'cat test-data.txt | ./build/bin/spicy-driver t.hlto' 'cat test-data.txt | ./build/bin/spicy-driver t-compiled-with-realpath.hlto'
Benchmark 1: cat test-data.txt | ./build/bin/spicy-driver t.hlto
  Time (mean ± σ):      1.218 s ±  0.008 s    [User: 1.204 s, System: 0.029 s]
  Range (min … max):    1.209 s …  1.232 s    10 runs
 
Benchmark 2: cat test-data.txt | ./build/bin/spicy-driver t-compiled-with-realpath.hlto
  Time (mean ± σ):      1.219 s ±  0.005 s    [User: 1.202 s, System: 0.031 s]
  Range (min … max):    1.215 s …  1.231 s    10 runs
 
Summary
  cat test-data.txt | ./build/bin/spicy-driver t.hlto ran
    1.00 ± 0.01 times faster than cat test-data.txt | ./build/bin/spicy-driver t-compiled-with-realpath.hlto
```

Closes #1589.
Closes #1591.