@bbannier 2023-11-10
The remaining work here is dealing with the fallout from type mismatches from `const char*` vs `std::string`. This is likely exactly why we introduced the explicit typing. I picked `const char*` instead of `std::string_view` since `const char*` can implicitly convert to `std::string`, but maybe introducing a dedicated type would give us more control.