@bbannier 2023-11-18
I slightly reworked this since while we can emit C++ literals for generated (effectively static) strings user-specified literals should still be emitted as mutable strings (e.g., so users can call mutating functions on them). The way I now detect the difference is to check whether a literal has a location.