@bbannier 2023-11-20
I added a couple of other perf cleanups to this PR, please let me know if I should pull them into their own PR.